# openapi.json
{
  "components": {
    "schemas": {
...
      "CurrentRoutinesResponse": {
        "description": "Active routine prompts at the server's current time.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RoutinePromptResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "server_time",
          "items"
        ],
        "title": "CurrentRoutinesResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RoutineCategory": {
        "description": "Supported routine categories in the MVP.",
        "enum": [
          "MEAL",
          "MEDICATION"
        ],
        "title": "RoutineCategory",
        "type": "string"
      },
      "RoutineExecutionResponse": {
        "description": "Persisted result for one routine occurrence.",
        "properties": {
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "confirmation_delay_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Delay Seconds"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reminder_count": {
            "title": "Reminder Count",
            "type": "integer"
          },
          "routine_id": {
            "title": "Routine Id",
            "type": "string"
          },
          "scheduled_at": {
            "format": "date-time",
            "title": "Scheduled At",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RoutineState"
          }
        },
        "required": [
          "execution_id",
          "routine_id",
          "name",
          "state",
          "scheduled_at",
          "reminder_count",
          "confirmed_at",
          "confirmation_delay_seconds",
          "closed_at"
        ],
        "title": "RoutineExecutionResponse",
        "type": "object"
      },
      "RoutinePromptResponse": {
        "description": "One active routine prompt rendered and spoken by the tablet.",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/RoutineCategory"
          },
          "confirm_label": {
            "title": "Confirm Label",
            "type": "string"
          },
          "display_text": {
            "title": "Display Text",
            "type": "string"
          },
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reminder_count": {
            "title": "Reminder Count",
            "type": "integer"
          },
          "routine_id": {
            "title": "Routine Id",
            "type": "string"
          },
          "scheduled_at": {
            "format": "date-time",
            "title": "Scheduled At",
            "type": "string"
          },
          "spoken_text": {
            "title": "Spoken Text",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RoutineState"
          }
        },
        "required": [
          "execution_id",
          "routine_id",
          "name",
          "category",
          "state",
          "scheduled_at",
          "reminder_count",
          "display_text",
          "spoken_text",
          "confirm_label"
        ],
        "title": "RoutinePromptResponse",
        "type": "object"
      },
      "RoutineState": {
        "description": "Lifecycle states for one scheduled routine occurrence.",
        "enum": [
          "REMINDING",
          "CONFIRMED",
          "NOT_ANSWERED"
        ],
        "title": "RoutineState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "GuardianSession": {
        "in": "cookie",
        "name": "reminiscence_guardian_session",
        "type": "apiKey"
      },
      "TabletSession": {
        "in": "cookie",
        "name": "reminiscence_tablet_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Reminiscence API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/routines/current": {
      "get": {
        "description": "Advance due transitions and return prompts the tablet should display.",
        "operationId": "get_current_routines_api_v1_routines_current_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentRoutinesResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "TabletSession": []
          }
        ],
        "summary": "Get active tablet routine prompts",
        "tags": [
          "routines"
        ]
      }
    },
    "/api/v1/routines/history": {
      "get": {
        "description": "Advance transitions and return routine history without prompt text.",
        "operationId": "get_routine_history_api_v1_routines_history_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoutineExecutionResponse"
                  },
                  "title": "Response Get Routine History Api V1 Routines History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "GuardianSession": []
          }
        ],
        "summary": "List persisted routine executions",
        "tags": [
          "routines"
        ]
      }
    },
    "/api/v1/routines/{execution_id}/confirm": {
      "post": {
        "description": "Record the tablet button press using the trusted server time.",
        "operationId": "confirm_routine_api_v1_routines__execution_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineExecutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "TabletSession": []
          }
        ],
        "summary": "Confirm a completed routine",
        "tags": [
          "routines"
        ]
      }
    },
...
  }
}